t149: split Redux store (3,095 lines) into domain slices
Closes #733
